+++ dakaraphi.devtools.tracing/src/main/java/dakaraphi/devtools/tracing/ClassMethodSelector.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import dakaraphi.devtools.tracing.config.Tracer;
 import dakaraphi.devtools.tracing.config.TracingConfig;
+import dakaraphi.devtools.tracing.config.Tracer.Type;
 
 public class ClassMethodSelector {
 	public static final int LINE_LAST = -1;
@@ -30,16 +31,16 @@ public class ClassMethodSelector {
     	return false;
 	}
     
-    public boolean shouldTransformMethod(final String classNameDotted, final String methodName) {
-		Tracer definition = findMatchingDefinition(classNameDotted, methodName);
+    public boolean shouldTransformMethod(final String classNameDotted, final String methodName, final List<String> typeNames) {
+		Tracer definition = findMatchingDefinition(classNameDotted, methodName, typeNames);
     	if (definition != null && definition.enabled)
     		return true;
     	return false;
     }
 	
-	public Tracer findMatchingDefinition(final String classNameDotted, final String methodName) {
+	public Tracer findMatchingDefinition(final String classNameDotted, final String methodName, final List<String> typeNames) {
 		for (final Tracer definition : definitionList) {
-			if ( doesDefinitionMatch(definition, classNameDotted, methodName)) {
+			if ( doesDefinitionMatch(definition, classNameDotted, methodName, typeNames)) {
 				return definition;
 			}
 		}
@@ -47,7 +48,7 @@ public class ClassMethodSelector {
 	}
 	public Tracer findEnabledMatchingDefinition(final String classNameDotted, final String methodName) {
 		for (final Tracer definition : definitionList) {
-			if ( doesDefinitionMatch(definition, classNameDotted, methodName) && definition.enabled) {
+			if ( doesDefinitionMatch(definition, classNameDotted, methodName, null) && definition.enabled) {
 				return definition;
 			}
 		}
@@ -58,10 +59,23 @@ public class ClassMethodSelector {
 	 * Find definition that matches the classname and methodname or
 	 * just classname if methodname is null.
 	 */
-	private boolean doesDefinitionMatch(final Tracer definition, final String classNameDotted, final String methodName) {
-		if ( definition.classRegex.matcher(classNameDotted).matches() && (methodName == null || definition.methodRegex.matcher(methodName).matches())) {
+	private boolean doesDefinitionMatch(final Tracer definition, final String classNameDotted, final String methodName, final List<String> typeNames) {
+		if ( definition.classRegex.matcher(classNameDotted).matches() 
+		     && (methodName == null || definition.methodRegex.matcher(methodName).matches())  // if methodName null then match not required on method
+			 && (typeNames == null || doTypesMatch(definition, typeNames))) { // if typeNames null then match not required on types
 			return true;
 		}
 		return false;
+	}
+
+	private boolean doTypesMatch(final Tracer definition, final List<String> typeNames) {
+		for (Type type : definition.types) {
+			if (((type.index < typeNames.size()) && (type.typeRegex.matcher(typeNames.get(type.index)).matches()))) {
+				// all specified definitions for types must match
+				continue;
+			}
+			return false;
+		}
+		return true;
 	}
 }