+++ dakaraphi.devtools.tracing/src/main/java/dakaraphi/devtools/tracing/TracingAgent.java
@@ -33,7 +33,7 @@ import dakaraphi.devtools.tracing.metrics.ExecutionCounts;
 public class TracingAgent {
 	public static TracingConfig tracingConfig = null;
     public static void premain(String agentArgs, Instrumentation instrumentation) throws IOException, ClassNotFoundException, NoSuchMethodException, SecurityException, IllegalAccessException, IllegalArgumentException, InvocationTargetException, InstantiationException {
-    	TraceLogger.log("Starting v1.0.7");
+    	TraceLogger.log("Starting v1.0.8");
 		String tracerDefinitionFile = System.getProperty(ConfigurationSerializer.FILE_PROPERTY_KEY);
 		if (tracerDefinitionFile == null) {
 			TraceLogger.log("Missing system property " + ConfigurationSerializer.FILE_PROPERTY_KEY);
@@ -77,5 +77,7 @@ public class TracingAgent {
  - prevent infinite recursion if someone attempts to trace a JDK class that the tracer itself uses
  - allow tracers to override the global logConfig
  - add support for timers.  Just measure and log diff between each occurrence of timer by same name. option, on same thread.
+ - tracing a low level function like System.out.println.  The class will have already been loaded when the agent starts and therefore will not invoke the transformer
+	- it will get invoked if the tracer file is touched.  Need to investigate if there is a way to detect at agent startup.
 */