+++ dakaraphi.devtools.tracing/src/test/java/dakaraphi/devtools/TestApplication.java
@@ -32,6 +32,4 @@ public class TestApplication {
 
         }
     }
-
-
 }